pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.13",
    "better-profanity>=0.7.0",
    "discord-py>=2.5.2",
    "langdetect>=1.0.9",
]
